+++ CHANGELOG.md
@@ -9,8 +9,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - Ask iOS permission for "Critical Alert"
 - Validate OS permission before Android 13
-- Send acknowledge to remote server from Android
+- Send acknowledge to remote server on Android
 - Create notification when iOS app received background push notifications and app is inactive
+- Send acknowledge to remote server on IOS
 
 ## [6.0.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/push-notifications@6.0.0...@capacitor/push-notifications@6.0.1) (2024-06-13)